backend/djangoAPI/API/requirements.txt
asgiref==3.2.10
backcall==0.2.0
blis==0.4.1
boto3==1.16.9
botocore==1.19.9
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.3
Cython==0.29.14
decorator==4.4.2
dill==0.3.2
Django==3.1.2
django-rest-swagger==2.2.0
djangorestframework==3.12.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
filelock==3.0.12
gensim==3.8.3
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
importlib-metadata==2.0.0
ipython==7.18.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
langdetect==1.0.8
MarkupSafe==1.1.1
murmurhash==1.0.2
mysqlclient==2.0.1
nlp==0.4.0
numpy==1.19.2
openapi-codec==1.3.2
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.3
parso==0.7.1
pathy==0.3.1
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
preshed==3.0.2
prompt-toolkit==3.0.8
protobuf==3.13.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6.1
pyenchant==3.1.1
Pygments==2.7.2
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytokenizations==0.6.0
pytz==2020.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.3
simplejson==3.17.2
six==1.15.0
smart-open==2.2.1
spacy==2.3.2
sqlparse==0.4.1
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.51.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
wasabi==0.8.0
wcwidth==0.2.5
xxhash==2.0.0
zipp==3.4.0
